DCAP attestation changes 1. Add DCAP attestation support for Avalon. 2. It support DCAP Quote generation, quote verification in enclave manager side and Quote verification on client side. 3. Changes to generic client to usage of client side quote verification.

+++ tc/sgx/trusted_worker_manager/enclave_untrusted/enclave_bridge_wrapper/singleton/signup_info_singleton.cpp
@@ -97,11 +97,11 @@ std::map<std::string, std::string> SignupInfoSingleton::UnsealEnclaveData() {
 }  // SignupInfoSingleton::UnsealEnclaveData
 
 // XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX
-size_t SignupInfoSingleton::VerifyEnclaveInfo(const std::string& enclave_info,
+int SignupInfoSingleton::VerifyEnclaveInfo(const std::string& enclave_info,
     const std::string& mr_enclave) {
 
     SignupDataSingleton signup_data;
     tcf_err_t result = signup_data.VerifyEnclaveInfo(
         enclave_info, mr_enclave);
-    return (size_t) result;
+    return (int) result;
 }  // SignupInfoSingleton::VerifyEnclaveInfo
